improve style, fix some typos

+++ app/src/main/java/com/example/fict/MyMarkerView.java
@@ -1,4 +1,3 @@
-
 package com.example.fict;
 
 import android.annotation.SuppressLint;
@@ -24,7 +23,6 @@ public class MyMarkerView extends MarkerView {
 
     public MyMarkerView(Context context, int layoutResource) {
         super(context, layoutResource);
-
         tvContent = findViewById(R.id.tvContent);
     }
 
@@ -32,22 +30,17 @@ public class MyMarkerView extends MarkerView {
     // content (user-interface)
     @Override
     public void refreshContent(Entry e, Highlight highlight) {
-
         if (e instanceof CandleEntry) {
-
             CandleEntry ce = (CandleEntry) e;
-
             tvContent.setText(Utils.formatNumber(ce.getHigh(), 0, true));
         } else {
-
             tvContent.setText(Utils.formatNumber(e.getY(), 0, true));
         }
-
         super.refreshContent(e, highlight);
     }
 
     @Override
     public MPPointF getOffset() {
-        return new MPPointF(-(getWidth() / 2), -getHeight());
+        return new MPPointF(-(getWidth() / (float) 2), -getHeight());
     }
 }
